TD CTO Version 1.21 - Revised 9/7/2016 Prashant

+++ tcm/modules/web/src/main/java/com/invessence/web/dao/common/UserInfoDAO.java
@@ -29,7 +29,7 @@ public class UserInfoDAO extends JdbcDaoSupport
 
       Map outMap = sp.addUser(action, userData);
 
-      return ((Long) outMap.get(WebConst.LOGONID_PARAM)).longValue();
+      return ((Long) outMap.get(WebConst.LOGONID_PARAM));
    }
 
    public Boolean selectUserInfo(UserData data)
